do not dynamically allocate selection.{classic,edge}
